Separated logo and text, added startup loader, and updated all instances of the loader
