+++ package.json
@@ -15,6 +15,7 @@
     "@supabase/supabase-js": "^2.39.7",
     "clsx": "^2.1.0",
     "date-fns": "^3.3.1",
+    "framer-motion": "^12.4.7",
     "lucide-react": "^0.344.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
